+++ src/cli_gen/lib/src/code/models/runner_model.dart
@@ -7,5 +7,6 @@ class RunnerModel extends SubcommandModel {
     required super.docComments,
     required super.userClassName,
     required super.annotations,
+    required super.bound,
   });
 }